refactor: move thread runtime owner reads
